+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/GameSessionService.cs
@@ -379,21 +379,32 @@ namespace Stormancer.Server.Plugins.GameSession
             }
 
             var client = new Client(peer);
-
-            if (!_clients.TryAdd(user, client))
+            lock (_clients)
             {
-                if (_clients.TryGetValue(user, out Client alreadyConnectedClient) && alreadyConnectedClient.Status != PlayerStatus.Disconnected && !_clients.TryUpdate(user, client, alreadyConnectedClient))
+                if (!_clients.TryAdd(user, client))
                 {
-                    throw new ClientException("Failed to add player to the game session.");
+                    if (_clients.TryGetValue(user, out var alreadyConnectedClient))
+                    {
+                        if (alreadyConnectedClient.Status != PlayerStatus.Disconnected)
+                        {
+                            throw new ClientException("User already connected to gamesession.");
+                        }
+                        else if (!_clients.TryUpdate(user, client, alreadyConnectedClient))
+                        {
+                            throw new ClientException("Failed to update peer associated with user.");
+                        }
+                    }
                 }
             }
         }
 
         private async Task PeerConnectionRejected(IScenePeerClient peer)
         {
-            var client = _clients.FirstOrDefault(kvp => kvp.Value.Peer == peer);
-            _clients.TryRemove(client.Key, out _);
-
+            lock (_clients)
+            {
+                var client = _clients.FirstOrDefault(kvp => kvp.Value.Peer == peer);
+                _clients.TryRemove(client.Key, out _);
+            }
             await Task.CompletedTask;
         }
 
@@ -708,24 +719,25 @@ namespace Stormancer.Server.Plugins.GameSession
             _analytics.Push("gamesession", "playerLeft", JObject.FromObject(new { sessionId = peer.SessionId, gameSessionId = this._scene.Id }));
             Client client = null;
             string userId = null;
-
-            // the peer disconnected from the app and is not in the sessions anymore.
-            foreach (var kvp in _clients)
+            lock (_clients)
             {
-                if (kvp.Value.Peer == peer)
+                // the peer disconnected from the app and is not in the sessions anymore.
+                foreach (var kvp in _clients)
                 {
-                    userId = kvp.Key;
-                    client = kvp.Value;
-
-                    if (_config.Public)
+                    if (kvp.Value.Peer == peer)
                     {
-                        _clients.TryRemove(userId, out client);
+                        userId = kvp.Key;
+                        client = kvp.Value;
+
+                        if (_config.Public)
+                        {
+                            _clients.TryRemove(userId, out client);
+                        }
+                        // no need to continue searching for the client, we already found it
+                        break;
                     }
-                    // no need to continue searching for the client, we already found it
-                    break;
                 }
             }
-
             if (client != null)
             {
                 client.Peer = null;