+++ src/Stormancer.Plugins/Profile/Stormancer.Server.Plugins.Profile/ProfilePluginConstants.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Stormancer.Server.Plugins.Profile
+{
+    public static class ProfilePluginConstants
+    {
+        public const string SCENE_ID = "profiles";
+        public const string SCENE_TEMPLATE = "profiles";
+        public const string SERVICE_ID = "stormancer.profiles";
+    }
+}