+++ src/Stormancer.Plugins/Profile/Stormancer.Server.Plugins.Profile/Changelog.rst
@@ -11,9 +11,11 @@ This project adheres to semantic versioning.
 
 Unreleased
 ----------
-Changed
-*******
-- Automatically create scene and registration in service locator.
+Added
+*****
+- Add ICustomProfilePart interface to easily provides update/delete API for profiles using the ProfileController Update/Delete APIs.
+- Add CustomPartAttribute to declare class as profile parts.
+- Automatically create the profile scene on startup and adds service locator configuration.
 
 2.0.1
 -----